+++ Editor/AssetImportProcessor.cs
@@ -2,6 +2,8 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
+using System.Reflection;
+using NUnit.Framework;
 using UnityEditor;
 using UnityEditor.Experimental.AssetImporters;
 using UnityEditor.Presets;